gx10-mcp/pyproject.toml
[project]
name = "gx10-mcp"
version = "0.1.0"
description = "GX10 MCP Server — infrastructure management for ASUS Ascent GX10"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=3.1",
    "redis>=5.0",
    "pyyaml>=6.0",
]

[tool.uv]
dev-dependencies = []
